test: AccessTokenManager 테스트 작성
